requirements.txt
keras==3.1.1
numpy==1.26.4
pillow==10.2.0
streamlit==1.32.2
tensorflow==2.16.1